requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
django-admin-tailwind==0.0.31
django-appconf==1.0.6
django-ckeditor==6.7.0
django-compressor==4.4
django-js-asset==2.2.0
django-role-permissions==3.2.0
Pillow==10.1.0
python-decouple==3.8
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
typing-extensions==4.9.0
